requirements.txt
pandas
python-dotenv
requests>=2.25
pytest
coverage
Pylama
Sphinx
sphinx_rtd_theme
sphinxcontrib-programoutput
pycodestyle
pylint
beautifulsoup4